Skip unnecessary API calls
The callback was not guaranteed first try, hence there was a need to wait until the first successful callback which meant that multiple UDFs were posting WM_CLOSE and interrupting calculation before the ```m_calculationInProgress``` flag was set to True. However, adding the extra check guarantees the callback to work first try so there is no need to wait and so the very first UDF will definitely trigger the callback even if the user presses End/Continue/Debug on the 'code execution was interrupted' dialog.